refactor: 커스텀 예외 적용
