+++ src/main/java/woowacourse/shoppingcart/application/CartService.java
@@ -9,8 +9,8 @@ import woowacourse.shoppingcart.dao.CustomerDao;
 import woowacourse.shoppingcart.dao.ProductDao;
 import woowacourse.shoppingcart.domain.Cart;
 import woowacourse.shoppingcart.domain.Product;
-import woowacourse.shoppingcart.exception.InvalidProductException;
-import woowacourse.shoppingcart.exception.NotInCustomerCartItemException;
+import woowacourse.shoppingcart.exception.notfound.InvalidProductException;
+import woowacourse.shoppingcart.exception.notfound.NotInCustomerCartItemException;
 
 @Service
 @Transactional(rollbackFor = Exception.class)
