+++ src/main/java/woowacourse/shoppingcart/application/CustomerService.java
@@ -6,7 +6,7 @@ import woowacourse.shoppingcart.dao.CustomerDao;
 import woowacourse.shoppingcart.domain.Customer;
 import woowacourse.shoppingcart.dto.CustomerCreationRequest;
 import woowacourse.shoppingcart.dto.CustomerUpdationRequest;
-import woowacourse.shoppingcart.exception.DuplicateEmailException;
+import woowacourse.shoppingcart.exception.badrequest.DuplicateEmailException;
 
 @Service
 @Transactional(rollbackFor = Exception.class)
