+++ src/main/java/woowacourse/shoppingcart/dao/CustomerDao.java
@@ -9,8 +9,8 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.stereotype.Repository;
 import woowacourse.shoppingcart.domain.Customer;
-import woowacourse.shoppingcart.exception.InvalidCustomerException;
-import woowacourse.shoppingcart.exception.NotFoundCustomerException;
+import woowacourse.shoppingcart.exception.notfound.InvalidCustomerException;
+import woowacourse.shoppingcart.exception.notfound.NotFoundCustomerException;
 
 @Repository
 public class CustomerDao {
