--- src/main/java/woowacourse/shoppingcart/exception/DuplicateEmailException.java
@@ -1,8 +1,0 @@
-package woowacourse.shoppingcart.exception;
-
-public class DuplicateEmailException extends IllegalRequestException {
-
-    public DuplicateEmailException() {
-        super("1001", "이메일이 중복입니다.");
-    }
-}
