--- src/main/java/woowacourse/shoppingcart/exception/IllegalRequestException.java
@@ -1,10 +1,0 @@
-package woowacourse.shoppingcart.exception;
-
-import org.springframework.http.HttpStatus;
-
-public class IllegalRequestException extends ShoppingCartException {
-
-    public IllegalRequestException(final String errorCode, final String message) {
-        super(errorCode, message, HttpStatus.BAD_REQUEST);
-    }
-}
