--- src/main/java/woowacourse/shoppingcart/exception/InvalidLoginException.java
@@ -1,8 +1,0 @@
-package woowacourse.shoppingcart.exception;
-
-public class InvalidLoginException extends IllegalRequestException {
-
-    public InvalidLoginException() {
-        super("1002", "로그인에 실패했습니다.");
-    }
-}
