+++ src/main/java/woowacourse/shoppingcart/exception/badrequest/InvalidLoginException.java
@@ -1,0 +1,8 @@
+package woowacourse.shoppingcart.exception.badrequest;
+
+public class InvalidLoginException extends NotFoundException {
+
+    public InvalidLoginException() {
+        super("1002", "로그인에 실패했습니다.");
+    }
+}
