+++ src/main/java/woowacourse/shoppingcart/exception/notfound/InvalidCustomerException.java
@@ -1,0 +1,10 @@
+package woowacourse.shoppingcart.exception.notfound;
+
+import woowacourse.shoppingcart.exception.badrequest.NotFoundException;
+
+public class InvalidCustomerException extends NotFoundException {
+
+    public InvalidCustomerException() {
+        super("1004", "존재하지 않는 유저입니다.");
+    }
+}
