+++ src/main/java/woowacourse/shoppingcart/exception/notfound/NotFoundException.java
@@ -1,6 +1,7 @@
-package woowacourse.shoppingcart.exception;
+package woowacourse.shoppingcart.exception.notfound;
 
 import org.springframework.http.HttpStatus;
+import woowacourse.shoppingcart.exception.ShoppingCartException;
 
 public class NotFoundException extends ShoppingCartException {
