+++ src/main/java/woowacourse/shoppingcart/exception/unauthorized/UnauthorizedTokenException.java
@@ -1,6 +1,7 @@
-package woowacourse.shoppingcart.exception;
+package woowacourse.shoppingcart.exception.unauthorized;
 
 import org.springframework.http.HttpStatus;
+import woowacourse.shoppingcart.exception.ShoppingCartException;
 
 public class UnauthorizedTokenException extends ShoppingCartException {
