+++ src/test/java/woowacourse/shoppingcart/application/CustomerServiceIntegrationTest.java
@@ -15,8 +15,8 @@ import woowacourse.shoppingcart.dao.CustomerDao;
 import woowacourse.shoppingcart.domain.Customer;
 import woowacourse.shoppingcart.dto.CustomerCreationRequest;
 import woowacourse.shoppingcart.dto.CustomerUpdationRequest;
-import woowacourse.shoppingcart.exception.DuplicateEmailException;
-import woowacourse.shoppingcart.exception.NotFoundCustomerException;
+import woowacourse.shoppingcart.exception.badrequest.DuplicateEmailException;
+import woowacourse.shoppingcart.exception.notfound.NotFoundCustomerException;
 
 @SpringBootTest
 @Transactional
